requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic[email]==2.6.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.24.1
websockets==11.0.3
scholarly==1.7.11
beautifulsoup4==4.12.2
aiohttp==3.9.3
python-dotenv==1.0.0
boto3==1.34.49
supabase==1.0.4
PyJWT==2.8.0
selenium==4.12.0
webdriver-manager==4.0.1
pytest==7.4.2
pytest-asyncio==0.21.1
loguru==0.7.0 